+++ CHANGELOG.md
@@ -4,5 +4,6 @@
 
 ### Added
 
+- Switch for nomad job #9
 - Initial draft #2
 - Changelog #6